test: move Redis e2e fixture into utils
